require config.js => JSON.parse(fs.readSync config.js) in main();
require does not find config.js in the current working directory, but instead
uses the source directory.  This is not useful behavior for running
multiple simulations from various directories and is fixed by moving to
JSON.parse/fs.readSync.

As this primarily affects stand-alone nodejs usage, and not usage from
a browser, there is no associated testing.