make logging optional.  If this.options.logs is undefined, then these default logs are enabled:
        this.logs.trade  = new Log("./trades.csv");
        this.logs.order  = new Log('./orders.csv');
        this.logs.profit = new Log('./profit.csv');
        this.logs.ohlc   = new Log('./ohlc.csv');
        this.logs.volume = new Log('./volume.csv');

Otherwise, the keys in this.options.logs with a truthy value define which logs are created.

Also, if statements guard each log.write, so any this.logs entry can be deleted at any time and logging to that log will stop
with no error thrown.

Untested at this time, but not expected to cause issues.

Previous tests pass when this.options.logs is blank